Remove circular require
