+++ lib/exifr/tiff.rb
@@ -1,6 +1,6 @@
 # Copyright (c) 2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015 - R.W. van 't Veer
 
-require 'exifr'
+require 'exifr/common'
 require 'rational'
 require 'enumerator'
